+++ app/scripts/backend/stats-proxy.ts
@@ -4,7 +4,7 @@ import settings from "@/settings"
 import BaseProxy from "@/backend/base-proxy"
 import type { AjaxSettings, KorpResponse, ProgressResponse, ProgressCallback } from "@/backend/types"
 import { StatsNormalized, StatsColumn, StatisticsWorkerResult } from "@/statistics.types"
-import { locationSearchGet, httpConfAddMethod } from "@/util"
+import { locationSearchGet, httpConfAddMethod, Factory } from "@/util"
 import { statisticsService } from "@/statistics"
 
 /**
@@ -25,7 +25,7 @@ export function normalizeStatsData(data: KorpStatsResponse): StatsNormalized {
     return { ...data, combined, corpora }
 }
 
-export default class StatsProxy extends BaseProxy {
+export class StatsProxy extends BaseProxy {
     prevParams: KorpStatsParams | null
     prevRequest: AjaxSettings | null
     prevUrl?: string
@@ -158,6 +158,9 @@ export default class StatsProxy extends BaseProxy {
         return def.promise()
     }
 }
+
+const statsProxyFactory = new Factory(StatsProxy)
+export default statsProxyFactory
 
 /** @see https://ws.spraakbanken.gu.se/docs/korp#tag/Statistics/paths/~1count/get */
 type KorpStatsParams = {