+++ app/scripts/parallel/stats_proxy.ts
@@ -1,11 +1,13 @@
 /** @format */
-import model from "@/backend/model"
+import statsProxyFactory, { StatsProxy } from "@/backend/stats-proxy"
 import settings from "@/settings"
 
-model.StatsProxy = class ParallelStatsProxy extends model.StatsProxy {
+class ParallelStatsProxy extends StatsProxy {
     makeParameters(reduceVals: string[], cqp: string, ignoreCase: boolean) {
         let params = super.makeParameters(reduceVals, cqp, ignoreCase)
         params.within = settings.corpusListing.getAttributeQuery("within").replace(/\|.*?:/g, ":")
         return params
     }
 }
+
+statsProxyFactory.setClass(ParallelStatsProxy)