+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 - Improved UI reactivity for Simple search
 - Make the search button(s) more visible [#308](https://github.com/spraakbanken/korp-frontend/issues/308)
 - Use native checkboxes in corpus chooser, not images [#362](https://github.com/spraakbanken/korp-frontend/issues/362)
-- Added TypeScript definitions for Korp backend parameters and responses, and moved `@/model` to `@/backend/model`
+- Added TypeScript definitions for Korp backend parameters and responses
+- Wrapped `GraphProxy`, `KwicProxy`, `LemgramProxy`, `StatsProxy` and `TimeProxy` with factories; see [@/util/Factory](./app/scripts/util.ts)
 - Removed globals – import them instead (or their members):
   - `authenticationProxy`
   - `settings`